requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
numpy==2.2.1
python-dotenv==1.0.1
httpx==0.28.1
